Fix local 3rd party static issues (CodeQL) (#377)

+++ 3rd/fstrm/libmy/argv.c
@@ -1227,8 +1227,7 @@ static	int	string_to_value(const char *arg, ARGV_PNT var,
     }
     else if (arr_p->aa_entry_n % ARRAY_INCR == 0) {
       arr_p->aa_entries =
-	(char *)realloc(arr_p->aa_entries, (arr_p->aa_entry_n + ARRAY_INCR) *
-			size);
+	(char *)realloc(arr_p->aa_entries, (size_t)(arr_p->aa_entry_n + ARRAY_INCR) * size);
     }
     
     if (arr_p->aa_entries == NULL) {
